requirements.txt
httpx>=0.27,<1
cryptography>=42,<44
python-dotenv>=1,<2
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
jinja2>=3.1,<4
aiosqlite>=0.20,<1
